code/requirements.txt
pillow==11.1.0
pyglet==2.1.3
tk==0.1.0